React: refatorate timer component; separate into minor components
