+++ src/main.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import './css/style.css';
 import { Header } from './app/Header';
-import { Timer } from './app/Timer';
+import { Timer } from './app/Timer/Timer';
 import { Footer } from './app/Footer';
 
 ReactDOM.createRoot(document.getElementById('root')).render(
